REFACTOR: Melhorando os padrões de arquitetura dos metodos getPdf, DownloadPdf e FilesToZip
